exceptions left

+++ include/polynomial.hpp
@@ -19,7 +19,7 @@ class Polynomial
 		float			_a, _b, _c, _y1, _y2;
 		std::string		_sign;
 		int				_pos;
-		int 			_discriminant;
+		float 			_discriminant;
 
 		std::vector<std::string> _unsortedTerms;
 		std::vector<std::string> _sortedTerms;
@@ -28,7 +28,7 @@ class Polynomial
 		void	checkSignOfFirstTerm(void);
 		void	checkSignAfterEqual(int pos);
 		void	removeSpaces(void);
-		int		solveDiscriminant(std::vector<std::string> v, bool flag);
+		float	solveDiscriminant(std::vector<std::string> v, bool flag);
 		void	getDiscriminant(std::string degree);
 		std::vector<std::string>	getTheSameDegreeTerms(std::vector<std::string> temp);
 		void	displayDiscriminant(void);
